+++ v4/adw/adw-toolbar-view.go
@@ -247,7 +247,7 @@ func (x *ToolbarView) AddTopBar(WidgetVar *gtk.Widget) {
 	xToolbarViewAddTopBar(x.GoPointer(), WidgetVar.GoPointer())
 }
 
-var xToolbarViewGetBottomBarHeight func(uintptr) int32
+var xToolbarViewGetBottomBarHeight func(uintptr) int
 
 // Gets the current bottom bar height for @self.
 //
@@ -255,7 +255,7 @@ var xToolbarViewGetBottomBarHeight func(uintptr) int32
 // [property@ToolbarView:reveal-bottom-bars], including during the transition.
 //
 // See [method@ToolbarView.get_top_bar_height].
-func (x *ToolbarView) GetBottomBarHeight() int32 {
+func (x *ToolbarView) GetBottomBarHeight() int {
 	cret := xToolbarViewGetBottomBarHeight(x.GoPointer())
 	return cret
 }
@@ -317,7 +317,7 @@ func (x *ToolbarView) GetRevealTopBars() bool {
 	return cret
 }
 
-var xToolbarViewGetTopBarHeight func(uintptr) int32
+var xToolbarViewGetTopBarHeight func(uintptr) int
 
 // Gets the current top bar height for @self.
 //
@@ -325,7 +325,7 @@ var xToolbarViewGetTopBarHeight func(uintptr) int32
 // [property@ToolbarView:reveal-top-bars], including during the transition.
 //
 // See [method@ToolbarView.get_bottom_bar_height].
-func (x *ToolbarView) GetTopBarHeight() int32 {
+func (x *ToolbarView) GetTopBarHeight() int {
 	cret := xToolbarViewGetTopBarHeight(x.GoPointer())
 	return cret
 }
@@ -484,10 +484,10 @@ func (c *ToolbarView) SetGoPointer(ptr uintptr) {
 // [property@ToolbarView:reveal-bottom-bars], including during the transition.
 //
 // See [property@ToolbarView:top-bar-height].
-func (x *ToolbarView) GetPropertyBottomBarHeight() int32 {
+func (x *ToolbarView) GetPropertyBottomBarHeight() int {
 	var v gobject.Value
 	x.GetProperty("bottom-bar-height", &v)
-	return v.GetLong()
+	return v.GetInt()
 }
 
 // SetPropertyExtendContentToBottomEdge sets the "extend-content-to-bottom-edge" property.
@@ -619,10 +619,10 @@ func (x *ToolbarView) GetPropertyRevealTopBars() bool {
 // including during the transition.
 //
 // See [property@ToolbarView:bottom-bar-height].
-func (x *ToolbarView) GetPropertyTopBarHeight() int32 {
+func (x *ToolbarView) GetPropertyTopBarHeight() int {
 	var v gobject.Value
 	x.GetProperty("top-bar-height", &v)
-	return v.GetLong()
+	return v.GetInt()
 }
 
 // Requests the user's screen reader to announce the given message.
@@ -691,7 +691,7 @@ func (x *ToolbarView) GetAtContext() *gtk.ATContext {
 // This functionality can be overridden by `GtkAccessible`
 // implementations, e.g. to get the bounds from an ignored
 // child widget.
-func (x *ToolbarView) GetBounds(XVar *int32, YVar *int32, WidthVar *int32, HeightVar *int32) bool {
+func (x *ToolbarView) GetBounds(XVar *int, YVar *int, WidthVar *int, HeightVar *int) bool {
 	cret := gtk.XGtkAccessibleGetBounds(x.GoPointer(), XVar, YVar, WidthVar, HeightVar)
 	return cret
 }
@@ -807,7 +807,7 @@ func (x *ToolbarView) UpdateProperty(FirstPropertyVar gtk.AccessibleProperty, va
 // property change must be communicated to assistive technologies.
 //
 // This function is meant to be used by language bindings.
-func (x *ToolbarView) UpdatePropertyValue(NPropertiesVar int32, PropertiesVar []gtk.AccessibleProperty, ValuesVar []gobject.Value) {
+func (x *ToolbarView) UpdatePropertyValue(NPropertiesVar int, PropertiesVar []gtk.AccessibleProperty, ValuesVar []gobject.Value) {
 	gtk.XGtkAccessibleUpdatePropertyValue(x.GoPointer(), NPropertiesVar, PropertiesVar, ValuesVar)
 }
 
@@ -839,7 +839,7 @@ func (x *ToolbarView) UpdateRelation(FirstRelationVar gtk.AccessibleRelation, va
 // relation change must be communicated to assistive technologies.
 //
 // This function is meant to be used by language bindings.
-func (x *ToolbarView) UpdateRelationValue(NRelationsVar int32, RelationsVar []gtk.AccessibleRelation, ValuesVar []gobject.Value) {
+func (x *ToolbarView) UpdateRelationValue(NRelationsVar int, RelationsVar []gtk.AccessibleRelation, ValuesVar []gobject.Value) {
 	gtk.XGtkAccessibleUpdateRelationValue(x.GoPointer(), NRelationsVar, RelationsVar, ValuesVar)
 }
 
@@ -872,7 +872,7 @@ func (x *ToolbarView) UpdateState(FirstStateVar gtk.AccessibleState, varArgs ...
 // state change must be communicated to assistive technologies.
 //
 // This function is meant to be used by language bindings.
-func (x *ToolbarView) UpdateStateValue(NStatesVar int32, StatesVar []gtk.AccessibleState, ValuesVar []gobject.Value) {
+func (x *ToolbarView) UpdateStateValue(NStatesVar int, StatesVar []gtk.AccessibleState, ValuesVar []gobject.Value) {
 	gtk.XGtkAccessibleUpdateStateValue(x.GoPointer(), NStatesVar, StatesVar, ValuesVar)
 }
